Fix state_id truncation bug causing silent patterns
StateInitData::state_id was uint32_t while Instruction::state_id was
uint16_t. This caused init_seq_step() to create state at a different
key than SEQ_STEP opcode looked up, resulting in empty state and no
triggers firing.

Changed all state_id types to uint16_t consistently across:
- StateInitData, StatePool, VM, ProgramSlot, and WASM bindings

+++ akkado/include/akkado/codegen.hpp
@@ -14,7 +14,7 @@ namespace akkado {
 
 /// State initialization data for SEQ_STEP and TIMELINE opcodes
 struct StateInitData {
-    std::uint32_t state_id;
+    std::uint16_t state_id;  // Must match Instruction::state_id (16-bit)
     enum class Type : std::uint8_t {
         SeqStep,   // Initialize SeqStepState with timed events
         Timeline   // Initialize TimelineState with breakpoints
